prova

+++ src/main/java/com/tbr/lettura/repository/UserChallengeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.tbr.lettura.model.UserChallenge;
 
-public interface UserChallengeRepository extends JpaRepository<UserChallenge, Long> {
+public interface UserChallengeRepository extends JpaRepository<UserChallenge, Integer> {
     List<UserChallenge> findByUserId(Integer userId);
     UserChallenge findByUserIdAndChallengeId(Integer userId, Integer challengeId);
 }
